Allow faceted surfaces to output to .vtu files.
Add the member functions output_paraview() (Consistent with the
corresponding member function of oomph::Mesh) to TetMeshFacetedSurface.

The first output_paraview function takes in a std::ostream to output to,
this function contains all the relevant code.
The second output_paraview function takes in a filename, creates the
relevant std::ostream and runs the first instance of the output_paraview
function.

The function output_paraview may be further optimised since the function
loops over every vertex on every facet first to add them to a vector of
vertices.
Then does the same loop again to set the values of connectivity, offsets
and types. This may be reformulated as a single loop over the vertices
if more efficiency is required.